rename Command to LocalProcess
