+++ Tests/formicTests/CommandTests.swift
@@ -4,7 +4,7 @@ import Testing
 
 @Test("initializing a shell command")
 func shellCommandDeclarationTest() async throws {
-    let command = Command.shell("uname")
+    let command = LocalProcess.shell("uname")
     #expect(command.retry == .none)
     #expect(command.args == ["uname"])
     #expect(command.env == nil)
@@ -15,14 +15,14 @@ func shellCommandDeclarationTest() async throws {
 
 @Test("unique command ids by instance (Identifiable)")
 func verifyIdentifiableCommands() async throws {
-    let command1 = Command.shell("uname")
-    let command2 = Command.shell("uname")
+    let command1 = LocalProcess.shell("uname")
+    let command2 = LocalProcess.shell("uname")
     try #require(command1.id != command2.id)
 }
 
 @Test("initializing a shell command with all options")
 func shellCommandFullDeclarationTest() async throws {
-    let command = Command.shell(
+    let command = LocalProcess.shell(
         "ls", env: ["PATH": "/usr/bin"],
         retry: .retryOnFailure(
             Backoff(maxRetries: 200, strategy: .exponential(maxDelay: .seconds(60)))))
@@ -39,7 +39,7 @@ func shellCommandFullDeclarationTest() async throws {
 
 @Test("initializing a copy command")
 func copyCommandDeclarationTest() async throws {
-    let command = Command.remoteCopy(from: "one", to: "two")
+    let command = LocalProcess.remoteCopy(from: "one", to: "two")
     #expect(command.retry == .none)
     #expect(command.args == ["one", "two"])
     #expect(command.env == nil)
@@ -49,7 +49,7 @@ func copyCommandDeclarationTest() async throws {
 
 @Test("initializing a copy command with all options")
 func copyCommandFullDeclarationTest() async throws {
-    let command = Command.remoteCopy(
+    let command = LocalProcess.remoteCopy(
         from: "one", to: "two",
         retry: .retryOnFailure(Backoff(maxRetries: 100, strategy: .fibonacci(maxDelay: .seconds(60)))))
     #expect(command.args == ["one", "two"])
